BR 2.4 FA-L01-00-0174: Rule definition changes to allow CF=0 or CF>=1

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/rules/service/business/AbstractFact.java
@@ -13,6 +13,11 @@
 
 package eu.europa.ec.fisheries.uvms.rules.service.business;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.util.*;
 import com.google.common.base.Predicates;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
@@ -34,12 +39,6 @@ import org.joda.time.DateTimeZone;
 import un.unece.uncefact.data.standard.reusableaggregatebusinessinformationentity._20.*;
 import un.unece.uncefact.data.standard.unqualifieddatatype._20.DateTimeType;
 import un.unece.uncefact.data.standard.unqualifieddatatype._20.TextType;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.*;
 
 @Slf4j
 @ToString
@@ -852,7 +851,7 @@ public abstract class AbstractFact {
         ImmutableList<NumericType> removeNull = ImmutableList.copyOf(Iterables.filter(numericList, Predicates.notNull()));
         for (NumericType type : removeNull) {
             BigDecimal val = type.getValue();
-            equals = val.intValue() == value;
+            equals = val.doubleValue() == (double) value;
         }
         return equals;
     }
